dataplane: Don't include DPDK header in dataplane API header
