+++ src/cmds/Makefile.in
@@ -1,10 +1,6 @@
 TOPDIR		= @TOPDIR@
 MKRULESDIR	= @MKRULESDIR@
 
-RTE_SDK		= @RTE_SDK@
-RTE_TARGET	= @RTE_TARGET@
-RTE_LIBS	= @RTE_LIBS@
-
 TARGET_EXE		= lagopus
 INSTALL_SBIN_EXE_TARGETS	= $(TARGET_EXE)
 INSTALL_SBIN_EXE_DIR		= $(DEST_SBIN_EXEDIR)
@@ -19,9 +15,6 @@ DEP_LIBS	+= $(DEP_LAGOPUS_DATAPLANE_LIB)
 DEP_LIBS	+= $(DEP_LAGOPUS_AGENT_LIB)
 DEP_LIBS	+= $(DEP_LAGOPUS_DATASTORE_LIB)
 DEP_LIBS	+= $(DEP_LAGOPUS_UTIL_LIB)
-ifneq ($(RTE_SDK),)
-DEP_LIBS	+= $(DEP_DPDK_LIB)
-endif
 
 DEP_LIBS	+= @DEP_SNMP_HANDLER_LIB@
 LDFLAGS		+= @NETSNMP_AGENT_LIBS@
