+++ src/dataplane/dpdk/worker.c
@@ -100,6 +100,7 @@
 #include "lagopus/flowdb.h"
 #include "lagopus/meter.h"
 #include "lagopus/port.h"
+#include "lagopus/interface.h"
 #include "lagopus_gstate.h"
 #include "lagopus/ofp_dp_apis.h"
 #include "lagopus/ofcache.h"
@@ -109,6 +110,7 @@
 #include "pktbuf.h"
 #include "packet.h"
 #include "csum.h"
+#include "lock.h"
 #include "dpdk/dpdk.h"
 
 #ifndef APP_LCORE_WORKER_FLUSH
@@ -142,6 +144,7 @@ app_lcore_worker(struct app_lcore_params_worker *lp,
   static const uint8_t eth_bcast[] = { 0xff, 0xff, 0xff, 0xff, 0xff, 0xff };
   struct port *port;
   struct lagopus_packet *pkt;
+  enum switch_mode mode;
   uint32_t i;
 
 
@@ -202,7 +205,8 @@ app_lcore_worker(struct app_lcore_params_worker *lp,
       dp_interface_send_packet_kernel(pkt, pkt->in_port->interface);
 #endif /* HYBRID */
 
-      if (port->bridge->flowdb->switch_mode == SWITCH_MODE_STANDALONE) {
+      flowdb_switch_mode_get(port->bridge->flowdb, &mode);
+      if (mode == SWITCH_MODE_STANDALONE) {
         lagopus_forward_packet_to_port(pkt, OFPP_NORMAL);
         lp->mbuf_in.array[j] = NULL;
         continue;
