+++ src/dataplane/ofproto/datapath.c
@@ -38,6 +38,7 @@
 #include "lagopus/flowdb.h"
 #include "lagopus/bridge.h"
 #include "lagopus/port.h"
+#include "lagopus/interface.h"
 #include "lagopus/group.h"
 #include "lagopus/meter.h"
 #include "lagopus/dataplane.h"
@@ -1341,6 +1342,39 @@ packet_in_free(struct eventq_data *data) {
     }
     free(data);
   }
+}
+
+int
+lagopus_send_packet_physical(struct lagopus_packet *pkt,
+                             struct interface *ifp) {
+  if (ifp == NULL) {
+    return LAGOPUS_RESULT_OK;
+  }
+  switch (ifp->info.type) {
+    case DATASTORE_INTERFACE_TYPE_ETHERNET_DPDK_PHY:
+    case DATASTORE_INTERFACE_TYPE_ETHERNET_DPDK_VDEV:
+#ifdef HAVE_DPDK
+      return dpdk_send_packet_physical(pkt, ifp);
+#else
+      break;
+#endif
+    case DATASTORE_INTERFACE_TYPE_ETHERNET_RAWSOCK:
+      return rawsock_send_packet_physical(pkt,
+                                          ifp->info.eth_rawsock.port_number);
+
+    case DATASTORE_INTERFACE_TYPE_GRE:
+    case DATASTORE_INTERFACE_TYPE_NVGRE:
+    case DATASTORE_INTERFACE_TYPE_VXLAN:
+    case DATASTORE_INTERFACE_TYPE_VHOST_USER:
+    case DATASTORE_INTERFACE_TYPE_UNKNOWN:
+      /* TODO */
+      lagopus_packet_free(pkt);
+      return LAGOPUS_RESULT_OK;
+    default:
+      break;
+  }
+
+  return LAGOPUS_RESULT_INVALID_ARGS;
 }
 
 static lagopus_result_t
